+++ src/fiuba/algo3/tp2/HachaDeMetal.java
@@ -1,12 +1,12 @@
 package fiuba.algo3.tp2;
 
-import fiuba.algo3.tp2.estrategiasDesgaste.EstrategiaDesgasteDivididaDeFuerza;
+import fiuba.algo3.tp2.estrategiasDesgaste.EstrategiaDesgasteLineal;
 
 public class HachaDeMetal extends Hacha {
 
     public HachaDeMetal() {
         this.durabilidad = 400;
         this.fuerza = 10;
-        this.desgastador = new EstrategiaDesgasteDivididaDeFuerza(this.durabilidad, fuerza, 2);
+        this.desgastador = new EstrategiaDesgasteLineal(this.durabilidad, fuerza, 2);
     }
 }
