+++ src/fiuba/algo3/tp2/PicoPiedra.java
@@ -1,11 +1,11 @@
 package fiuba.algo3.tp2;
 
-import fiuba.algo3.tp2.estrategiasDesgaste.EstrategiaDesgasteDivididaDeFuerza;
+import fiuba.algo3.tp2.estrategiasDesgaste.EstrategiaDesgasteLineal;
 
 public class PicoPiedra extends Pico {
     public PicoPiedra() {
         this.durabilidad = 200;
         this.fuerza = 4;
-        this.desgastador = new EstrategiaDesgasteDivididaDeFuerza(this.durabilidad, this.fuerza, 1.5);
+        this.desgastador = new EstrategiaDesgasteLineal(this.durabilidad, this.fuerza, 1.5);
     }
 }
