--- src/fiuba/algo3/tp2/estrategiasDesgaste/EstrategiaDesgasteDivididaDeFuerza.java
@@ -1,20 +1,0 @@
-package fiuba.algo3.tp2.estrategiasDesgaste;
-
-public class EstrategiaDesgasteDivididaDeFuerza extends EstrategiaDesgaste {
-
-    private double divisor;
-
-    public EstrategiaDesgasteDivididaDeFuerza(int durabilidad, int fuerza, double divisor){
-
-        this.durabilidad = durabilidad;
-        this.fuerza = fuerza;
-        this.divisor = divisor;
-    }
-
-
-    @Override
-    public void desgastar(){
-
-        this.durabilidad -= (this.fuerza / this.divisor);
-    }
-}
