+++ test/fiuba/algo3/tp2/EstrategiaDesgasteLinealTest.java
@@ -12,13 +12,13 @@ public class EstrategiaDesgasteLinealTest {
         int durabilidadInical = 10;
         int fuerza = 3;
 
-        EstrategiaDesgasteLineal desgaste = new EstrategiaDesgasteLineal(durabilidadInical, fuerza);
+        EstrategiaDesgasteLineal desgaste = new EstrategiaDesgasteLineal(durabilidadInical, fuerza,1);
 
         Assert.assertEquals(durabilidadInical, desgaste.durabilidad());
     }
 
     @Test
-    public void desgasteLinealReduceSegunFuerzaAsiganda(){
+    public void desgasteLinealNoRecibeDivisorEntoncesReduceSegunFuerzaAsiganda(){
 
         int durabilidadInical = 10;
         int fuerza = 3;
@@ -31,4 +31,41 @@ public class EstrategiaDesgasteLinealTest {
 
         Assert.assertEquals(durabilidadActual, desgaste.durabilidad());
     }
+
+    @Test
+    public void desgasteDivididaPorDivisorReduceSegunFuerzaYDivisorAsigando() {
+
+        int durabilidadInical = 10;
+        int fuerza = 3;
+        double divisor = 2;
+        int durabilidadActual = durabilidadInical;
+
+        EstrategiaDesgasteLineal desgaste = new EstrategiaDesgasteLineal(durabilidadInical, fuerza, divisor);
+
+        desgaste.desgastar();
+        durabilidadActual -= fuerza / divisor;
+
+        Assert.assertEquals(durabilidadActual, desgaste.durabilidad());
+    }
+
+
+    @Test
+    public void desgasteDivididaPorDivisorReduceSegunFuerzaYDivisorAsigando2() {
+
+        /* el divisor es numero con decimal */
+
+        int durabilidadInical = 10;
+        int fuerza = 3;
+        double divisor = 1.5;
+        int durabilidadActual = durabilidadInical;
+
+        EstrategiaDesgasteLineal desgaste = new EstrategiaDesgasteLineal(durabilidadInical, fuerza, divisor);
+
+        desgaste.desgastar();
+        durabilidadActual -= fuerza / divisor;
+
+        Assert.assertEquals(durabilidadActual, desgaste.durabilidad());
+    }
+
+
 }
